+++ src/libs/balutil/inc/balinfo.h
@@ -29,6 +29,13 @@ typedef enum BAL_INFO_PACKAGE_TYPE
     BAL_INFO_PACKAGE_TYPE_BUNDLE_PATCH,
 } BAL_INFO_PACKAGE_TYPE;
 
+typedef enum BAL_INFO_CACHE_TYPE
+{
+    BAL_INFO_CACHE_TYPE_NO,
+    BAL_INFO_CACHE_TYPE_YES,
+    BAL_INFO_CACHE_TYPE_ALWAYS,
+} BAL_INFO_CACHE_TYPE;
+
 
 typedef struct _BAL_INFO_PACKAGE
 {
@@ -43,7 +50,7 @@ typedef struct _BAL_INFO_PACKAGE
     LPWSTR sczUpgradeCode;
     LPWSTR sczVersion;
     LPWSTR sczInstallCondition;
-    BOOL fAlwaysCache;
+    BAL_INFO_CACHE_TYPE cacheType;
 } BAL_INFO_PACKAGE;